requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-chroma
chromadb
pymupdf
pytesseract
pillow
python-multipart
python-dotenv
google-genai
fastembed
numpy
PyJWT
email-validator
bcrypt
